+++ packages/frameworks/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @vercel/frameworks
 
+## 3.8.4
+
+### Patch Changes
+
+- This updates `Nuxt.js` -> `Nuxt`, detects Nuxt if the nightly version is installed, and updates the development command to nuxt dev. ([#13606](https://github.com/vercel/vercel/pull/13606))
+
+## 3.8.3
+
+### Patch Changes
+
+- Adds FastAPI to the list of auto-detected frameworks ([#13877](https://github.com/vercel/vercel/pull/13877))
+
+## 3.8.2
+
+### Patch Changes
+
+- Add .cts support to express and hono builders ([#13828](https://github.com/vercel/vercel/pull/13828))
+
+- - Expand framework detection to src/app and src/server files. ([#13828](https://github.com/vercel/vercel/pull/13828))
+  - Improve handling when multiple entrypoints are detected.
+
+## 3.8.1
+
+### Patch Changes
+
+- Add support for app.js as a server entrypoint ([#13798](https://github.com/vercel/vercel/pull/13798))
+
 ## 3.8.0
 
 ### Minor Changes